Big update

+++ StormWeb/Controllers/BranchController.cs
@@ -44,13 +44,16 @@ namespace StormWeb.Controllers
 
         [Authorize(Roles = "Super,BranchManager")]
         [HttpPost]
-        public ActionResult Create(Branch branch)
+        public ActionResult Create(Branch branch, FormCollection fc)
         {
+            branch.Address.Country_Id = Convert.ToInt32(fc["Country_Id"]);
+
             if (ModelState.IsValid)
             {
                 db.Branches.AddObject(branch);
                 db.SaveChanges();
                 LogHelper.writeToSystemLog(new string[] { CookieHelper.Username }, (CookieHelper.Username + " Added a new branch " + branch.Branch_Name), LogHelper.LOG_CREATE, LogHelper.SECTION_BRANCH);
+                NotificationHandler.setNotification(NotificationHandler.NOTY_SUCCESS, "Successfully Created a new branch" + branch.Branch_Name);
                 return RedirectToAction("Index");
             }
 
@@ -77,6 +80,9 @@ namespace StormWeb.Controllers
 
             branchmodel.addressTable = l;
 
+            List<Country> c = new List<Country>();
+            c.Add(branchmodel.branchTable.First().Address.Country);
+            branchmodel.countryTable = c;
             return View(branchmodel);
         }
 
@@ -88,19 +94,90 @@ namespace StormWeb.Controllers
             {
                 Branch branches = new Branch();
                 int branchid = Convert.ToInt32(collection["item.Branch_Id"]);
+                int addressId = Convert.ToInt32(collection["Address_Id"]);
+                int countryId = Convert.ToInt32(collection["Country_Id"]);
+                branches.Branch_Id = branchid;
+
+                branchmodel.branchTable = db.Branches.ToList().Where(x => x.Branch_Id == branchid).ToList();
+                branchmodel.addressTable = db.Addresses.ToList().Where(x => x.Address_Id == addressId).ToList();
+                branchmodel.countryTable = db.Countries.ToList().Where(x => x.Country_Id == countryId).ToList();
+
                 branches = db.Branches.Single(x => x.Branch_Id == branchid);
 
-                Address address = branches.Address;
+                Address address =branches.Address;
+                branches.Address= db.Addresses.Single(x => x.Address_Id == addressId);
+                address = branches.Address;
+
                 Country country = branches.Address.Country;
+                branches.Address.Country = db.Countries.Single(x => x.Country_Id == countryId);
+                country = branches.Address.Country;
 
+                if (collection["Address_Id"] == "" || collection["Address_Id"] == null)
+                {
+                    ModelState.AddModelError("Address", "Please Enter Your Address");
+                    return View(branchmodel);
+                }
+                if (collection["Branch_Name"] == "" || collection["Branch_Name"] == null)
+                {
+                    ModelState.AddModelError("Branch_Name", "Please Enter Your Branch Name");
+                    return View(branchmodel);
+                }
+                if (collection["Address_Name"] == "" || collection["Address_Name"] == null)
+                {
+                    ModelState.AddModelError("Address_Name", "Please Enter Your Address Name");
+                    return View(branchmodel);
+                }
+                if (collection["City"] == "" || collection["City"] == null)
+                {
+                    ModelState.AddModelError("City", "Please Enter Your City");
+                    return View(branchmodel);
+                }
+                if (collection["State"] == "" || collection["State"] == null)
+                {
+                    ModelState.AddModelError("State", "Please Enter Your State");
+                    return View(branchmodel);
+                }
+                //if (collection["Zipcode"] == "" || collection["Zipcode"] == null)
+                //{
+                //    ModelState.AddModelError("Zipcode", "Please Enter Your Zipcode");
+                //    return View(branchmodel);
+                //}
+                if (collection["Primary_Contact"] == "" || collection["Primary_Contact"] == null)
+                {
+                    ModelState.AddModelError("Primary_Contact", "Please Enter Your Primary Contact");
+                    return View(branchmodel);
+                }
+                if (collection["Seconday_Contact"] == "" || collection["Secondary_Contact"] == null)
+                {
+                    ModelState.AddModelError("Seconday_Contact", "Please Enter Your Seconday Contact");
+                    return View(branchmodel);
+                }
+                if (collection["Email"] == "" || collection["Email"] == null)
+                {
+                    ModelState.AddModelError("Email", "Please Enter Your Email");
+                    return View(branchmodel);
+                } 
 
+                string str = collection["Address_Name"];
 
-                //db.Branches.Attach(branch);
+                branches.Address.Address_Name = collection["Address_Name"];
+                branches.Address.City = collection["City"];
+                branches.Address.State = collection["State"];
+                branches.Address.Zipcode = Convert.ToInt32(collection["Zipcode"]);
+                branches.Branch_Name = collection["Branch_Name"];
+                branches.Primary_Contact = Convert.ToInt32(collection["Primary_Contact"]);
+                branches.Secondary_Contact = Convert.ToInt32(collection["Seconday_Contact"]);
+                branches.Email = collection["Email"];
+                 
+                db.ObjectStateManager.ChangeObjectState(address, EntityState.Modified);
+                db.ObjectStateManager.ChangeObjectState(country, EntityState.Modified);
                 db.ObjectStateManager.ChangeObjectState(branches, EntityState.Modified);
+
                 db.SaveChanges();
                 LogHelper.writeToSystemLog(new string[] { CookieHelper.Username }, (CookieHelper.Username + " Edited the branch " + branches.Branch_Name), LogHelper.LOG_UPDATE, LogHelper.SECTION_BRANCH);
-
+                NotificationHandler.setNotification(NotificationHandler.NOTY_SUCCESS, "Successfully Edited branch" + branches.Branch_Name);
                 return RedirectToAction("Index");
+
             }
 
             //ViewBag.Address_Id = new SelectList(db.Addresses, "Address_Id", "Address_Name", branch.Address_Id);
@@ -124,9 +201,47 @@ namespace StormWeb.Controllers
         [Authorize(Roles = "Super,BranchManager")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Branch branch = db.Branches.Single(b => b.Branch_Id == id);
+            Branch branch = db.Branches.SingleOrDefault(b => b.Branch_Id == id);
+            Appointment appointment = db.Appointments.SingleOrDefault(x => x.Branch_Id == id);
+            Case cases = db.Cases.SingleOrDefault(x => x.Branch_Id == id);
+            Branch_Staff branchStaff = db.Branch_Staff.SingleOrDefault(x => x.Branch_Id == id);
+            Client client = db.Clients.SingleOrDefault(x => x.Branch_Id == id);
+            General_Enquiry generalEnquiry = db.General_Enquiry.SingleOrDefault(x => x.Branch_Id == id);
+
+            if (appointment != null)
+            {
+                ModelState.AddModelError("Branch", "Branch " + branch.Branch_Name + " cannot be deleted Since there are Appointment's assigned to that branch");
+                NotificationHandler.setNotification(NotificationHandler.NOTY_WARNING, "Branch " + branch.Branch_Name + " cannot be deleted Since there are Appointment's assigned to that branch");
+                return View(branch);
+            }
+            if (cases != null)
+            {
+                ModelState.AddModelError("Branch", "Branch " + branch.Branch_Name + " cannot be deleted Since there are Case's assigned to that branch");
+                NotificationHandler.setNotification(NotificationHandler.NOTY_WARNING, "Branch " + branch.Branch_Name + " cannot be deleted Since there are Case's assigned to that branch");
+                return View(branch);
+            }
+            if (branchStaff != null)
+            {
+                ModelState.AddModelError("Branch", "Branch " + branch.Branch_Name + " cannot be deleted Since there are Staff's assigned to that branch");
+                NotificationHandler.setNotification(NotificationHandler.NOTY_WARNING, "Branch " + branch.Branch_Name + " cannot be deleted Since there are Staff's assigned to that branch");
+                return View(branch);
+            }
+            if (client != null)
+            {
+                ModelState.AddModelError("Branch", "Branch " + branch.Branch_Name + " cannot be deleted Since there are Client's assigned to that branch");
+                NotificationHandler.setNotification(NotificationHandler.NOTY_WARNING, "Branch " + branch.Branch_Name + " cannot be deleted Since there are Client's assigned to that branch");
+                return View(branch);
+            }
+            if (generalEnquiry != null)
+            {
+                ModelState.AddModelError("Branch", "Branch " + branch.Branch_Name + " cannot be deleted Since there are Enquiries related to that branch");
+                NotificationHandler.setNotification(NotificationHandler.NOTY_WARNING, "Branch " + branch.Branch_Name + " cannot be deleted Since there are Enquiries related to that branch");
+                return View(branch);
+            }
+
             db.Branches.DeleteObject(branch);
             db.SaveChanges();
+            NotificationHandler.setNotification(NotificationHandler.NOTY_SUCCESS, "Branch " + branch.Branch_Name + " is Deleted");
             LogHelper.writeToSystemLog(new string[] { CookieHelper.Username }, (CookieHelper.Username + " Deleted the branch " + branch.Branch_Name), LogHelper.LOG_DELETE, LogHelper.SECTION_BRANCH);
             return RedirectToAction("Index");
         }
